CT-122 블로그 게시글 작성 및 편집 페이지 리팩토링
